feat: api 로그인 유저정보

+++ Back/theartimposter/src/main/java/com/catpawdogpaw/theartimposter/security/api/controller/ReissueController.java
@@ -12,12 +12,18 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Controller;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.ResponseBody;
+import org.springframework.web.util.UriComponentsBuilder;
 
+import java.io.IOException;
 import java.util.Date;
+import java.util.HashMap;
+import java.util.Map;
 
 @Controller
 @ResponseBody
+@RequestMapping("/api")
 @RequiredArgsConstructor
 public class ReissueController {
 
@@ -25,7 +31,7 @@ public class ReissueController {
     private final RefreshRepository refreshRepository;
 
     @PostMapping("/reissue")
-    public ResponseEntity<?> reissue(HttpServletRequest request, HttpServletResponse response) {
+    public ResponseEntity<?> reissue(HttpServletRequest request, HttpServletResponse response) throws IOException {
 
         //get refresh token
         String refresh = null;
@@ -62,22 +68,31 @@ public class ReissueController {
             return new ResponseEntity<>("invalid refresh token", HttpStatus.BAD_REQUEST);
         }
 
+        //DB에 저장되어 있는지 확인
+        Boolean isExist = refreshRepository.existsByRefresh(refresh);
+        if (!isExist) {
+
+            //response body
+            return new ResponseEntity<>("invalid refresh token", HttpStatus.BAD_REQUEST);
+        }
+        String id = jwtUtil.getId(refresh);
         String username = jwtUtil.getUsername(refresh);
         String role = jwtUtil.getRole(refresh);
 
         //make new JWT
-        String newAccess = jwtUtil.createJwt("access", username, role, 600000L);
-        String newRefresh = jwtUtil.createJwt("refresh", username, role, 86400000L);
+        String newAccess = jwtUtil.createJwt("access", username, id, role, 600000L);
+        String newRefresh = jwtUtil.createJwt("refresh", username, id, role, 86400000L);
 
         //Refresh 토큰 저장 DB에 기존의 Refresh 토큰 삭제 후 새 Refresh 토큰 저장
         refreshRepository.deleteByRefresh(refresh);
         addRefreshEntity(username, newRefresh, 86400000L);
 
-        //response
-        response.setHeader("access", newAccess);
-        response.addCookie(createCookie("refresh", newRefresh));
+        //response as parameters
+        Map<String, String> tokens = new HashMap<>();
+        tokens.put("access", newAccess);
+        tokens.put("refresh", newRefresh);
 
-        return new ResponseEntity<>(HttpStatus.OK);
+        return new ResponseEntity<>(tokens, HttpStatus.OK);
     }
 
     private void addRefreshEntity(String username, String refresh, Long expiredMs) {
